chore(app): Remove the useless service worker (#162)
The service worker only cached the file `img/logo-agentj.png`, which is
not even used in the application.

However, a strange bug occured with Mail application on iOS after
retrieving an email: the user was redirected to this file. I don't even
know how it happened and even less how to reproduce the problem.

At least, with the file removed, there are less chances to be redirected
to it!