fix: Fix wblist import allowing bad values
The previous code allowed to import domains as well as emails. However,
the `FILTER_VALIDATE_DOMAIN` check wasn't strong enough and would allow
invalid data such as 'example' or '"example.com"' (with the quotes).

As the interface only mentions import of emails, we drop the import of
domains. The workaround is to add domains manually.