fix: Fix IMAP encryption selection
The documentation of PHP-IMAP is wrong about its own code. It says:

```
false — Disable encryption
ssl — Use SSL
tls — Use TLS
starttls — Use STARTTLS (alias TLS) (legacy only)
notls — Use NoTLS (legacy only)
```

In fact, SSL and TLS are merely the same thing, TLS being the
continuation of SSL. StartTLS on the other side is different thing as it
upgrades an insecure connection to SSL/TLS.

Besides the fact the documentation is wrong (starttls is NOT an alias
for TLS), the code in PHP-IMAP seems to be correct.