fix: Fix editing LDAP connector failure
The controller failed because the CryptEncrypt::decrypt method expects a
string while the password can be null.

I simply removed the line because `$oldPassClear` wasn't used. I guess
it was a typo in the `$oldPass !== $connector->getLdapPassword()` condition.

I changed the condition so the encrypted password is always set if the
password is filled, even if the password didn't changed: it's not a big
deal. However the form will need some refactoring to handle the password
properly.