fix(smtp): Remove default mailer From and Sender headers
When sending emails, Sender, From and Return-Path were set with the
default values. However, the From can be changed per domain (see the
mailAuthenticationSender parameter). The issue was than that Sender and
From were different for no reason.

As we always set the From header when sending mails, we don't need the
default values which generate some confusion.

Also as `DOMAIN_MAIL_AUTHENTIFICATION_SENDER` is generated on the 2.2
branch, an admin not satisfied with the default value would have to
change the value manually in the Docker container.