chore: Fix the failing CI
The CI was failing because, during the first test, the expected email
never arrived.

In fact, since 294476c8 and cd485617, the emails are sent asynchronously
by the "worker" Docker service. On the CI, the worker wasn't started
soon enough because it was started at the same time as the app service,
which didn't have time to execute the migrations. So the worker was
failing for a while and wasn't started during the first test.

Now, the worker waits for the app to be fully ready before being
started. It means that the migrations have been executed, so there is no
reason for the worker to fail.

Also, the tests script now waits for the worker to be healthy as well,
so we are always sure that everything is started before running the
tests.