fix: Fix setting the locale
Two problems happened. First, the `Accept-Language` header was ignored
when the user wasn't logged-in. It means that the French locale was
always used. The refactor of the LocaleSubscriber fixes this.

The second problem was that once the session expired, but that the user
was logged in again through the "remember me" cookie, the locale wasn't
set anymore. This is why I introduce a UserLocaleSubscriber which is
triggered on "interactive login", meaning it is called both on normal
login, and on "remember me" login.