fix: Fix autorelease with a high number of untreated messages
An issue happened when there was a high number of untreated messages
associated to users without the "bypass human authentication" option.
Indeed, the `getSearchQuery()` could return 1000 messages concerning
these users, while the 1001th message (not returned) could have
concerned a user with the "bypass" option enabled.

Now, we iterate on the users with the option enabled and we get their
untreated messages so we are sure to always load messages that need to
be loaded.